[Silver IV] Title: Trash Bins, Time: 716 ms, Memory: 299124 KB, Score: 11 point -BaekjoonHub

+++ 백준/Silver/23076. Trash Bins/Trash Bins.java
@@ -10,7 +10,7 @@ public class Main {
         for (int t = 0; t < T; t++) {
             int N = Integer.parseInt(BR.readLine());
             String S = BR.readLine();
-            int[] distance = new int[N];
+            long[] distance = new long[N];
             Arrays.fill(distance, N - 1);
             for (int i = 0; i < N;) {
                 if (S.charAt(i) == '1') {
